--- src/ui/editor/editors/ListEditor/index.ts
@@ -1,1 +1,0 @@
-import "./ListEditor.tsx"
