+++ src/ui/editor/editors/ObjectEditor/index.ts
@@ -1,1 +1,1 @@
-export { default as ObjectEditor } from './ObjectEditor.tsx'
+export * from './ObjectEditor.tsx'
